Improve generator for sumOp to consider the collection base type
Problem: The generator of sumOp always returned the result as a BigDecimal,
which is not obvious and a different behaviour compared to the interpreter.

Solution: The generator considers the collection base type. In case it's a NumberType,
it depends on the numbers precision whether the result will be a BigInteger or BigDecimal.

Because it's not possible to retrieve the generic type T of the PCollection<T> type in
former CollectionHelper.sum() (see javas type erasure), the different cases are handled via
separate reduction-rules.